dpif-hw-acc: fix netdev refcount and locks
Adding a port we had a reference to a netdev without increasing the
refcount, so when a netdev is closed by another thread, it was freed
and we had a invalid reference. Also we use a lock to protect a
race between getting a reference to it and removing it from
our datapath.

Issue: 895956
Change-Id: I14371f90c2c73f64e3efe6f5f73ce6f0b50eb5c3
Signed-off-by: Paul Blakey <paulb@mellanox.com>