tc: Add port mirroring offloading support
Add a new member ifindex_mirror in struct tc_flower to save the ifindex
of the mirroring port. If there are two output ports, the first one
is the mirroring port and the second one is the redirect port.
This patch parses this information and passes it to tc module to
offload.

Issue: 1173016
Change-Id: Ifb3c92a5283aa9a8a204eaffbe0144f74ef52151
Signed-off-by: Chris Mi <chrism@mellanox.com>