lib/netdev-tc-offloads: Fix frag first/later translation
Fragment mask (any and later) always exists so we need to test
for FLOW_NW_FRAG_LATER only if the state is FLOW_NW_FRAG_ANY.
Before this fix we could pass frag no and first at the same time to TC
which is also not tested there for bad frag state.
This fix make sure we only pass frag first/later if is frag.

Issue: 1315045
Fixes: 83e866067ea6 ("netdev-tc-offloads: Add support for IP fragmentation")
Change-Id: I2092d3d3197e0a0d0d7957d2949778e4af4274c2
Signed-off-by: Roi Dayan <roid@mellanox.com>