dpif-netdev: Initialize dpif_flow attrs
In a previous commit, "dpctl: Properly reflect a rule's offloaded
to HW state" (d63ca5329ff9), the dpif_flow struct was expanded, with
the 'offloaded' field being moved into a new struct which also
includes a field for the dp layer the flow is handled on. The
initialization of these fields was only done in dpif-netlink.

This completes that commit, by initializing the fields in dpif-netdev
as well. As the 'offloaded' field was previously ignored by
dpif-netdev, the attrs are initialized to the default values of
'false' for the offloaded state, and 'ovs' for the dp layer.

Issue: 1428539
Change-Id: I082c9f2b9267ef84b49ac30c82f1e445334e3bd4
Signed-off-by: Gavi Teitz <gavi@mellanox.com>